+++ docs/changelog.md
@@ -9,6 +9,9 @@
   - Add `box env --set-int KEY=VALUE` to add integer variables.
   - Add `box env --set-bool KEY=VALUE` to add boolean variables.
   - Add `box env --get VAR_NAME` to get the value of a variable.
+  - Add `box env --unset VAR_NAME` to remove a variable.
+  - Add `box env --list` to list all variables.
+- Bug fix for `box uninit`: Will throw a useful error if not in a `box` project.
 
 If this breaks your project, you can either run `box uninit` followed by `box init` and re-enter the variables, or you can manually edit the `pyproject.toml` file.